Store unsimplified items in Mongo
Earlier, orders were simplified and transformed before storing them in
Mongo. This removed a bunch of info that would be useful for improving
other commands, so we now store the unsimplified, complete data in
Mongo.

This also allows us to remove the orderizeItems transformer, which is
now done inline when transforming for perform.