Correct option prices
We now make an assumption that if an option is required and its price
matches the price of the item, it is a "default" option. This means that
selecting the option will not affect the price of the item, and will
hopefully result in correct subtotal calculations.

Also remove some extraneous waits.